Real­Time Restaurant Geolocation Service with Node.js
A simple real­time web application that will determine and show the
locations of the user’s input address directly on a map along with the
nearest restaurants extracted from a large CSV file located within a
particular radius. For this purpose I am using Node.js and MongoDB for
the server side and the HTML5 for front end. No​de.js​is an
asynchronous web server which is built on the Google V8 JavaScript
engine and is a perfect solution as a back­end for real­time apps. My
app will let users see addresses on the map with the help of the
MongoDB which is queried on the zipcode of the restaurants for a
real­time data channel. Our example will work in all modern browsers
that support the H​TML5.​